+++ lib/src/main/java/com/metaplex/lib/drivers/solana/AccountResponse.kt
@@ -9,20 +9,18 @@ package com.metaplex.lib.drivers.solana
 
 import com.metaplex.lib.experimental.serialization.serializers.base64.BorshAsBase64JsonArraySerializer
 import com.metaplex.lib.experimental.serialization.serializers.solana.AnchorAccountSerializer
+import com.metaplex.lib.experimental.serialization.serializers.solana.SolanaResponseSerializer
 import com.solana.models.buffer.Buffer
 import com.solana.models.buffer.BufferInfo
 import com.solana.vendor.borshj.BorshCodable
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.descriptors.SerialDescriptor
-import kotlinx.serialization.encoding.Decoder
-import kotlinx.serialization.encoding.Encoder
 
 @Serializable
 data class AccountInfo<D>(val data: D?, val executable: Boolean,
                           val lamports: Long, val owner: String?, val rentEpoch: Long)
 
-fun <D, T: BorshCodable> AccountInfo<D>.toBufferInfo() =
+internal fun <D, T: BorshCodable> AccountInfo<D>.toBufferInfo() =
     BufferInfo(data?.let { Buffer(data as T) }, executable, lamports, owner, rentEpoch)
 
 internal fun <A> SolanaAccountSerializer(serializer: KSerializer<A>) =
@@ -35,20 +33,8 @@ internal fun <A> SolanaAccountSerializer(serializer: KSerializer<A>) =
 internal inline fun <reified A> SolanaAccountSerializer() =
     AccountInfoSerializer<A?>(BorshAsBase64JsonArraySerializer(AnchorAccountSerializer()))
 
-internal class AccountInfoSerializer<D>(dataSerializer: KSerializer<D>)
-    : KSerializer<AccountInfo<D>?> {
-    private val serializer = WrappedValue.serializer(AccountInfo.serializer(dataSerializer))
-    override val descriptor: SerialDescriptor = serializer.descriptor
-
-    override fun serialize(encoder: Encoder, value: AccountInfo<D>?) =
-        encoder.encodeSerializableValue(serializer, WrappedValue(value))
-
-    override fun deserialize(decoder: Decoder): AccountInfo<D>? =
-        decoder.decodeSerializableValue(serializer).value
-}
-
-@Serializable
-private class WrappedValue<V>(val value: V?)
+private fun <D> AccountInfoSerializer(serializer: KSerializer<D>) =
+    SolanaResponseSerializer(AccountInfo.serializer(serializer))
 
 /*
  * The following code is alternative implementation that aliases the existing BufferInfo type
